Don't compile dependencies unless specified.
This is the second step in making builds idempotent. When dependencies
are compiled, there is always some sort of time-dependent thingy or
other squirrelled away in there. Accordingly we want to give users
the option to skip this step.

To override this, set `compile_dependencies: true` in your yaml file.